Consolidates and reorganizes code
