+++ ui/followedtags.go
@@ -1,6 +1,9 @@
 package ui
 
 import (
+	"context"
+	"strings"
+
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/container"
 	"fyne.io/fyne/v2/dialog"
@@ -60,11 +63,7 @@ func (ma *myApp) MakeFollowedTagsUI() fyne.CanvasObject {
 							"Tags successfully unfollowed",
 							ma.window).Show()
 					}
-					tags, err := GetFollowedTags(ma.prefs)
-					if err != nil {
-						dialog.NewError(err, ma.window)
-					}
-					ma.SetFollowedTags(tags)
+					ma.getFollowedTags()
 				}
 			}, ma.window)
 	})
@@ -84,4 +83,56 @@ func (ma *myApp) SetFollowedTags(t []*mastodon.FollowedTag) {
 	ma.listChoices.SetLeftItems(t)
 	ma.listChoices.SetRightItems([]*mastodon.FollowedTag{})
 	ma.listChoices.container.Refresh()
+}
+
+// RemoveFollowedTags removes the given list of tags from the user's following list.
+// These tags will no longer show up in the user's feed
+func (ma *myApp) RemoveFollowedTags(w fyne.Window) func() {
+	return func() {
+		dialog.ShowCustomConfirm("Confirm Unfollow", "Unfollow", "Cancel", makeRemoveConfirmUI(ma.listChoices.RightItems), func(b bool) {
+			if b {
+				c := NewClientFromPrefs(ma.prefs)
+				err := mastotool.RemoveFollowedTags(*c, ma.listChoices.RightItems)
+				if err != nil {
+					dialog.NewError(err, w)
+					return
+				} else {
+					tags := make([]string, len(ma.listChoices.RightItems))
+					for i, t := range ma.listChoices.RightItems {
+						tags[i] = t.Name
+					}
+					dialog.NewInformation(
+						"Success",
+						"Tags successfully unfollowed",
+						w).Show()
+				}
+				ma.getFollowedTags()
+			}
+		}, w)
+	}
+}
+
+// getFollowedTags gets the list of followed tags and populates the keepTags and removeTags based on this
+func (ma *myApp) getFollowedTags() {
+	c := NewClientFromPrefs(ma.prefs)
+	tags, err := c.GetFollowedTags(context.Background(), nil)
+	if err != nil {
+		tags = []*mastodon.FollowedTag{}
+		dialog.ShowError(err, ma.window)
+	}
+	ma.SetFollowedTags(tags)
+}
+
+// makeRemoveConfirmUI creates the UI for the tag removal confirmation action
+func makeRemoveConfirmUI(tags []*mastodon.FollowedTag) fyne.CanvasObject {
+	var sb strings.Builder
+	sb.WriteString("These tags will be removed from your following list and not seen in your feed:")
+	for _, t := range tags {
+		sb.WriteString("\n* " + t.Name)
+	}
+	rt := widget.NewRichTextFromMarkdown(sb.String())
+	rt.Wrapping = fyne.TextWrapWord
+	scroll := container.NewVScroll(rt)
+	scroll.SetMinSize(fyne.Size{Width: 100, Height: 150})
+	return scroll
 }
